refactor! : new feedback module + opaque types for libafl_setup
