+++ puffin/src/fuzzer/libafl_setup.rs
@@ -13,9 +13,7 @@ use super::harness;
 use crate::fuzzer::bit_mutations::{
     bit_mutations_dy, havoc_mutations_dy, MakeMessage, ReadMessage,
 };
-use crate::fuzzer::feedback::MinimizingFeedback;
-use crate::fuzzer::claim_feedback::ClaimFeedback;
-use crate::fuzzer::claim_observer::ClaimObserver;
+use crate::fuzzer::minimizer::MinimizingFeedback;
 use crate::fuzzer::mutations::{dy_mutations, MutationConfig};
 use crate::fuzzer::stages::PuffinMutationalStage;
 use crate::fuzzer::stats_monitor::StatsMonitor;
@@ -25,9 +23,8 @@ use crate::protocol::{ProtocolBehavior, ProtocolTypes};
 use crate::put::PutDescriptor;
 use crate::put_registry::PutRegistry;
 use crate::trace::{ConfigTrace, Spawner, Trace, TraceContext};
+use crate::fuzzer::feedback;
 
-pub const MAP_FEEDBACK_NAME: &str = "edges";
-const EDGES_OBSERVER_NAME: &str = "edges_observer";
 const MIN_BIT_EXECS: usize = 5_000; // one 1 core
 const MIN_BIT_CORPUS: usize = 200; // on 1 core
 
@@ -441,179 +438,6 @@ where
     }
 }
 
-type ConcreteMinimizer<S> = IndexesLenTimeMinimizerScheduler<QueueScheduler<S>>;
-
-type ConcreteObservers<'a> = (
-    HitcountsMapObserver<StdMapObserver<'a, u8, false>>,
-    (TimeObserver, (ClaimObserver, ())),
-);
-
-type ConcreteFeedback<'a, S> = CombinedFeedback<
-    MapFeedback<
-        DifferentIsNovel,
-        HitcountsMapObserver<StdMapObserver<'a, u8, false>>,
-        MaxReducer,
-        S,
-        u8,
-    >,
-    CombinedFeedback<
-        TimeFeedback,
-        ClaimFeedback<S>,
-        LogicEagerOr,
-        S,
-    >,
-    LogicEagerOr,
-    S,
->;
-
-impl<'harness, 'a, H, SC, C, R, EM, OF, CS, PT>
-    RunClientBuilder<
-        'harness,
-        H,
-        C,
-        R,
-        SC,
-        EM,
-        CombinedFeedback<
-            MinimizingFeedback<
-                StdState<
-                    Trace<PT>,
-                    CachedOnDiskCorpus<Trace<PT>>,
-                    RomuDuoJrRand,
-                    CachedOnDiskCorpus<Trace<PT>>,
-                >,
-                PT,
-            >,
-            CombinedFeedback<
-                MapFeedback<
-                    DifferentIsNovel,
-                    HitcountsMapObserver<StdMapObserver<'_, u8, false>>,
-                    MaxReducer,
-                    StdState<
-                        Trace<PT>,
-                        CachedOnDiskCorpus<Trace<PT>>,
-                        RomuDuoJrRand,
-                        CachedOnDiskCorpus<Trace<PT>>,
-                    >,
-                    u8,
-                >,
-                CombinedFeedback<
-                    TimeFeedback,
-                    ClaimFeedback<StdState<
-                            Trace<PT>, 
-                            CachedOnDiskCorpus<Trace<PT>>, 
-                            RomuDuoJrRand, 
-                            CachedOnDiskCorpus<Trace<PT>>
-                        >>,
-                    LogicEagerOr,
-                    StdState<Trace<PT>, CachedOnDiskCorpus<Trace<PT>>, RomuDuoJrRand, CachedOnDiskCorpus<Trace<PT>>>,
-                >,
-                LogicEagerOr,
-                StdState<
-                    Trace<PT>,
-                    CachedOnDiskCorpus<Trace<PT>>,
-                    RomuDuoJrRand,
-                    CachedOnDiskCorpus<Trace<PT>>,
-                >,
-            >,
-            LogicEagerOr,
-            StdState<
-                Trace<PT>,
-                CachedOnDiskCorpus<Trace<PT>>,
-                RomuDuoJrRand,
-                CachedOnDiskCorpus<Trace<PT>>,
-            >,
-        >,
-        OF,
-        ConcreteObservers<'a>,
-        CS,
-        PT,
-    >
-where
-    ConcreteState<C, R, SC, Trace<PT>>: UsesInput<Input = Trace<PT>>,
-    C: Corpus + UsesInput<Input = Trace<PT>> + fmt::Debug,
-    R: Rand,
-    SC: Corpus + UsesInput<Input = Trace<PT>> + fmt::Debug,
-    H: FnMut(&Trace<PT>) -> ExitKind,
-    OF: Feedback<ConcreteState<C, R, SC, Trace<PT>>>,
-    CS: Scheduler + UsesState<State = ConcreteState<C, R, SC, Trace<PT>>>,
-    EM: EventFirer
-        + EventRestarter
-        + EventManager<
-            ConcreteExecutor<
-                'harness,
-                H,
-                ConcreteObservers<'a>,
-                ConcreteState<C, R, SC, Trace<PT>>,
-            >,
-            StdFuzzer<
-                ConcreteMinimizer<ConcreteState<C, R, SC, Trace<PT>>>,
-                ConcreteFeedback<'a, ConcreteState<C, R, SC, Trace<PT>>>,
-                OF,
-                ConcreteObservers<'a>,
-            >,
-        > + ProgressReporter
-        + UsesState<State = ConcreteState<C, R, SC, Trace<PT>>>,
-    <EM as UsesState>::State: HasClientPerfMonitor + HasMetadata + HasExecutions,
-    PT: ProtocolTypes + 'static,
-{
-    fn create_feedback_observers(
-        &self,
-    )-> (
-        ConcreteFeedback<'a, ConcreteState<C, R, SC, Trace<PT>>>, 
-        ConcreteObservers<'a>,
-    )
-    {
-        #[cfg(not(test))]
-        let map = unsafe {
-            pub use libafl_targets::{EDGES_MAP, MAX_EDGES_NUM};
-            &mut EDGES_MAP[0..MAX_EDGES_NUM]
-        };
-
-        #[cfg(test)]
-        let map = unsafe {
-            // When testing we should not import libafl_targets, else it conflicts with sancov_dummy
-            pub const EDGES_MAP_SIZE: usize = 65536;
-            pub static mut EDGES_MAP: [u8; EDGES_MAP_SIZE] = [0; EDGES_MAP_SIZE];
-            pub static mut MAX_EDGES_NUM: usize = 0;
-            &mut EDGES_MAP[0..MAX_EDGES_NUM]
-        };
-
-        let map_feedback: MapFeedback<
-            DifferentIsNovel,
-            HitcountsMapObserver<StdMapObserver<'_, u8, false>>,
-            MaxReducer,
-            ConcreteState<C, R, SC, Trace<PT>>,
-            u8,
-        > = MaxMapFeedback::with_names_tracking(
-            MAP_FEEDBACK_NAME,
-            EDGES_OBSERVER_NAME,
-            true,
-            false,
-        );
-
-        {
-            let time_observer = TimeObserver::new("time");
-            let edges_observer =
-                HitcountsMapObserver::new(unsafe { StdMapObserver::new(EDGES_OBSERVER_NAME, map) });
-            let claim_observer = ClaimObserver::new("claim_observer");
-            let claim_feedback = ClaimFeedback::new();
-            let feedback = feedback_or!(
-                // New maximization map feedback linked to the edges observer and the feedback
-                // state `track_indexes` needed because of
-                // IndexesLenTimeMinimizerCorpusScheduler
-                map_feedback,
-                // Time feedback, this one does not need a feedback state
-                // needed for IndexesLenTimeMinimizerCorpusScheduler
-                TimeFeedback::with_observer(&time_observer),
-                claim_feedback  
-            );
-            let observers = tuple_list!(edges_observer, time_observer, claim_observer);
-            (feedback, observers)
-        }
-    }
-}
-
 /// Starts the fuzzing loop
 pub fn start<'harness, PB>(
     put_registry: &'harness PutRegistry<PB>,
@@ -698,7 +522,7 @@ where
         {
             // FIXME
             log::warn!("Running without minimizer is unsupported");
-            let (feedback, observer) = builder.create_feedback_observers();
+            let (feedback, observer) = feedback::build_feedback_and_observers();
             let feedback_with_minimizer = feedback_or!(
                 MinimizingFeedback::new(mutation_stage_config.with_truncation),
                 feedback
