+++ puffin/src/fuzzer/stats_monitor.rs
@@ -15,7 +15,7 @@ use libafl_bolts::prelude::*;
 use serde::Serialize;
 use serde_json::Serializer as JSONSerializer;
 
-use crate::fuzzer::libafl_setup::MAP_FEEDBACK_NAME;
+use crate::fuzzer::feedback::MAP_FEEDBACK_NAME;
 use crate::fuzzer::stats_stage::{RuntimeStats, STATS};
 
 trait ClonableMonitor: Monitor + DynClone {}
